Add expanded/floating music player with draggable scrubber and Android back button support
- Split music player into modular sub-components (bar, expanded, floating)
- Add full-screen expanded player overlay with pointer-event scrubber,
  volume, speed, share, and zap controls
- Add draggable floating pill that snaps to screen edges with drag-to-dismiss
- Add minimize (bar→floating) and expand (bar→expanded) buttons
- Mobile bar: remove scrubber/share/zap/volume/speed, add expand/minimize
- Desktop bar: keep all existing controls, add expand/minimize buttons
- Layout content offset adapts to player view mode (0px for floating)
- Android back button closes expanded player to bar mode
- Fix z-index conflicts: share modal z-70, poll creator z-80
- Fix lightbox safe-area padding for Android status bar
- Move Bible audio links into header dropdown menu
- Remove unused share buttons from note cards and threaded comments
